+++ packages/core/test/http-client/http-client.spec.ts
@@ -1,6 +1,7 @@
 import https from 'https';
 import Axios from 'axios';
 import nock from 'nock';
+import { createLogger } from '@sap-cloud-sdk/util';
 import {
   addDestinationToRequestConfig,
   buildHttpRequest,
@@ -70,6 +71,23 @@ describe('generic http client', () => {
 
       expect(actualProxy).toMatchObject(expectedProxy);
       expect(actualProxy.httpAgent).toBeDefined();
+    });
+
+    it('warn when custom headers are used', async () => {
+      const logger = createLogger({
+        package: 'core',
+        messageContext: 'http-client'
+      });
+      const warnSpy = jest.spyOn(logger, 'warn');
+
+      await buildHttpRequest(httpsDestination, {
+        authorization: 'abc',
+        'sap-client': '001',
+        'SAP-Connectivity-SCC-Location_ID': 'efg'
+      });
+      expect(warnSpy).toBeCalledWith(
+        'The custom headers are provided with the keys: authorization,sap-client,SAP-Connectivity-SCC-Location_ID. These keys will overwrite the headers created by the SDK.'
+      );
     });
 
     it('throws useful error messages when finding the destination fails', async () => {