+++ CHANGELOG.md
@@ -23,10 +23,12 @@
 ## Improvements
 
 - Remove the `rambda` library to save  approx 2MB or 10% of the SDK size.
+- [Connectivity] Warn the user when using custom headers for executing http requests.
 
 ## Fixed Issues
 
 - [Generator] Fix the function import (OData V2 + V4) and action import, where the return type is an primitive edm type like `Edm.String`.
+- [CSRF] Fix the error message of fetching the csrf token, it should show original error.
 
 
 # 1.30.0